fix(shell): rename RootContentInputGate to satisfy tsx naming
Architecture filename convention requires PascalCase / *-shell / *-ui
for .tsx files. The kebab-case gate file broke CI once the suite could
actually run after the checkout fix.
