+++ apps/cli/src/app-shell/RootContentInputGate.tsx
